Fix minor complaints

+++ src/tester.cpp
@@ -24,8 +24,8 @@ int convert_image(const ezsift::Image<unsigned char> &input,
   output->height = input.h;
   output->pixels = (float*)calloc(sizeof(float), input.w*input.h);
   if(output->pixels == 0) return 0;
-  for(size_t y=0; y<input.h; ++y){
-    for(size_t x=0; x<input.w; ++x){
+  for(int y=0; y<input.h; ++y){
+    for(int x=0; x<input.w; ++x){
       size_t idx = y*input.w+x;
       output->pixels[idx] = (float) input.data[idx];
     }
@@ -88,7 +88,7 @@ int run_tests(struct test *tests, uint32_t count){
   int passes = 0;
 
   fprintf(stderr, "\033[1;33m --> \033[0;0mRunning %i tests\n", count);
-  for(int i=0; i<count; ++i){
+  for(uint32_t i=0; i<count; ++i){
     if(run_test(tests[i])){
       passes++;
     }else{
